// view/dashboard/staff/MBE/types.ts
export type MBERecord = {
  Customers?: Array<{
    customerId?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    bvn?: string;
    dob?: string;
    inputtedDob?: string | null;
    dobMisMatch?: boolean;
    createdAt?: string;
    updatedAt?: string;
    customerLoanDiskId?: string;
    channel?: string;
    bvnPhoneNumber?: string;
    mainPhoneNumber?: string | null;
    mbeId?: string;
    monoCustomerConnectedCustomerId?: string;
    LoanRecord?: Array<{
      loanRecordId?: string;
      customerId?: string;
      loanDiskId?: string | null;
      customerLoanDiskId?: string;
      lastPoint?: string;
      channel?: string;
      loanStatus?: string;
      createdAt?: string;
      updatedAt?: string;
      loanAmount?: number;
      deviceId?: string | null;
      downPayment?: number;
      insurancePackage?: string | null;
      insurancePrice?: number;
      mbeEligibleAmount?: number;
      payFrequency?: string | null;
      storeId?: string | null;
      devicePrice?: number;
      deviceAmount?: number;
      monthlyRepayment?: number;
      duration?: number | null;
      interestAmount?: number;
      deviceName?: string | null;
      mbeId?: string | null;
    }>;
  }>;
  accountStatus?: string;
  bvn?: string | null;
  bvnPhoneNumber?: string | null;
  channel?: string;
  createdAt?: string;
  dob?: string | null;
  email?: string | null;
  firstname?: string;
  imagePublicId?: string | null;
  imageUrl?: string | null;
  isActive?: boolean;
  lastname?: string;
  mbeId?: string;
  mbe_old_id?: string;
  otp?: string | null;
  otpExpiry?: string | null;
  password?: string | null;
  phone?: string;
  resetOtp?: string | null;
  resetOtpExpiry?: string | null;
  role?: string;
  state?: string | null;
  stores?: Array<{
    id?: string;
    mbeId?: string | null;
    mbeOldId?: string;
    storeId?: string | null;
    storeOldId?: number;
  }>;
  title?: string;
  tokenVersion?: number;
  updatedAt?: string;
  username?: string;
};